CMR-10164: splitting out class to its own file.
