+++ subscription/build.sh
@@ -1,1 +1,1 @@
-zip deployment_package.zip subscription_worker.py part1_docker part_docker
+zip deployment_package.zip subscription_worker.py sns.py part1_docker part_docker
