+++ subscription/subscription_worker.py
@@ -1,9 +1,8 @@
 import boto3
-import json
 import multiprocessing
 import os
-import time
 from flask import Flask, jsonify
+from sns import Sns
 from sys import stdout
 from botocore.exceptions import ClientError
 
@@ -13,49 +12,6 @@ DEAD_LETTER_QUEUE_URL = os.getenv("DEAD_LETTER_QUEUE_URL")
 SUB_DEAD_LETTER_QUEUE_URL = os.getenv("SUB_DEAD_LETTER_QUEUE_URL")
 LONG_POLL_TIME = os.getenv("LONG_POLL_TIME")
 SNS_NAME = os.getenv("SNS_NAME")
-
-class sns:
-    """Encapsulates AWS SNS topics."""
-
-    def __init__(self, sns_resource):
-        """:param sns_resource: A Boto3 AWS SNS resource."""
-        self.sns_resource = sns_resource
-
-    def create_topic(self, topic_name):
-        """The topic is already created, but this is best practice to get the topic client
-        from AWS."""
-        try:
-            topic = self.sns_resource.create_topic(Name=topic_name)
-        except ClientError as error:
-            print("Count not get the topic ARN: {error}.")
-            stdout.flush()
-            raise error
-        else:
-            return topic
-
-    @staticmethod
-    def publish_message(topic, message):
-        """ Publishes a message with attributes to the CMR external topic. Subscriptions
-        can be filtered based on the message attributes. """
-        message_body_str = message["Body"]
-        message_body = json.loads(message_body_str)
-        message_subject = message_body["Subject"]
-        message_attributes = message_body["MessageAttributes"]
-        message_message = message_body["Message"]
-        try:
-            if message_attributes:
-                att_dict = {}
-                for key in message_attributes.keys():
-                    att_dict[key] = {"DataType": "String", "StringValue": message_attributes[key]["Value"]}
-                response = topic.publish(Subject=message_subject, Message=message_message, MessageAttributes=att_dict)
-            else:    
-                response = topic.publish(Subject=message_subject, Message=message_message)
-        except ClientError as error:
-            print(f"Could not publish message to topic {topic}. {error}")
-            stdout.flush()
-            raise error
-        else:
-            return response
 
 def receive_message(sqs_client, queue_url):
     response = sqs_client.receive_message(
@@ -116,7 +72,7 @@ def shutdown():
 
 sqs_client = boto3.client("sqs", region_name=AWS_REGION)
 sns_resource = boto3.resource("sns")
-sns_client = sns(sns_resource)
+sns_client = Sns(sns_resource)
 topic = sns_client.create_topic(SNS_NAME)
 
 #Shared boolean value for process communication
